Move medicalHistory component folders into components/ and fix imports
(cherry picked from commit 43a272ca1fe700153e1b57862dbba832f273379e)
