+++ frontend/src/modules/medicalHistory/components/MedicalTable/MedicalTable.script.js
@@ -1,5 +1,5 @@
 import { useMedicalTableStore } from '@/modules/medicalHistory/stores/medicalTableStore';
-import MnoChart from '@/modules/medicalHistory/MnoChart/MnoChart.vue';
+import MnoChart from '@/modules/medicalHistory/components/MnoChart/MnoChart.vue';
 import { buildIndicatorsFromRow } from '@/modules/shared/utils/vitalsHelpers';
 import { computed } from 'vue';
