+++ frontend/src/modules/medicalHistory/components/MedicalTable/MedicalTable.style.css
