+++ frontend/src/modules/medicalHistory/components/MedicalTable/MedicalTable.template.html
