+++ frontend/src/modules/medicalHistory/components/MedicalTable/MedicalTable.vue
