+++ frontend/src/modules/medicalHistory/components/MnoChart/MnoChart.script.js
