+++ frontend/src/modules/medicalHistory/components/MnoChart/MnoChart.style.css
