+++ frontend/src/modules/medicalHistory/components/MnoChart/MnoChart.template.html
