+++ frontend/src/modules/medicalHistory/components/MnoChart/MnoChart.vue
