+++ frontend/src/modules/medicalHistory/components/PatientCard/PatientCard.script.js
