+++ frontend/src/modules/medicalHistory/components/PatientCard/PatientCard.style.css
