+++ frontend/src/modules/medicalHistory/components/PatientCard/PatientCard.template.html
