+++ frontend/src/modules/medicalHistory/components/PatientCard/PatientCard.vue
