+++ frontend/src/modules/medicalHistory/components/Pharmacogenetics/Pharmacogenetics.script.js
