+++ frontend/src/modules/medicalHistory/components/Pharmacogenetics/Pharmacogenetics.style.css
