+++ frontend/src/modules/medicalHistory/components/Pharmacogenetics/Pharmacogenetics.template.html
