+++ frontend/src/modules/medicalHistory/components/Pharmacogenetics/Pharmacogenetics.vue
