+++ frontend/src/modules/medicalHistory/components/TreatmentCard/TreatmentCard.script.js
