+++ frontend/src/modules/medicalHistory/components/TreatmentCard/TreatmentCard.style.css
