+++ frontend/src/modules/medicalHistory/components/TreatmentCard/TreatmentCard.template.html
