+++ frontend/src/modules/medicalHistory/components/TreatmentCard/TreatmentCard.vue
